relocated qkd implementation from main.c to new file: qkd_management.c.
